cleans up uploader.js by concatting all js classes together from new folder: libraries
