+++ src/index.html
@@ -1508,6 +1508,37 @@
                         <h1>Uploader</h1>
                     </header>
 
+                    <div id="uploader_settings">
+                        <div>S3 Bucket Name <input type="text" class="uploader-setting" name="bucketName" /></div>
+                        <div>S3 Bucket Region <input type="text" class="uploader-setting" name="bucketRegion" /></div>
+                        <div>S3 IdentityPoolId <input type="text" class="uploader-setting" name="identityPoolId" /></div>
+                        <div>Server Address <input type="text" class="uploader-setting" name="serverAddress" /></div>
+
+                        <div><input type="button" value="Save Uploader Settings" id="save_uploader_settings"></div>
+                    </div>
+                    <script type="text/javascript">
+                        const settings = JSON.parse(localStorage.getItem('uploader-settings'));
+                        if (settings) {
+                            let settingFields = document.getElementsByClassName('uploader-setting');
+                            for (let i=0; i<settingFields.length; i++) {
+                                settingFields[i].value = Object.entries(settings)[i][1];
+                            }
+                        }
+
+                        document.getElementById('save_uploader_settings').addEventListener('click', function() {
+                            // get values of settings and store them
+                            const settings = document.getElementsByClassName('uploader-setting');
+                            let settingsObj = {};
+                            Array.from(settings).forEach(setting => {
+                                settingsObj[setting.name] = setting.value;
+                            });
+                            // save on localStorage
+                            localStorage.setItem('uploader-settings', JSON.stringify(settingsObj));
+                            // reload page so settings take effect
+                            location.reload()
+                        });
+                    </script>
+
                     <uniform-uploader id="file_uploader">
                     </uniform-uploader>
                     <div id="file_uploader_list">
@@ -1518,8 +1549,11 @@
                          const uploaderListDOM = document.querySelector('#file_uploader_list');
                          uploaderDOM.addEventListener('uploaded', (e) => {
                              console.log(e, e.detail);
-                             // add this to a variable
-                             uploaderListDOM.innerHTML = `<img src="${e.detail.Location}" />`;
+                             uploaderListDOM.innerHTML = "";
+
+                             for (let i=0; i<e.detail.length; i++) {
+                                 uploaderListDOM.innerHTML += `<img src="${e.detail[i].sourceUrl}" />`;
+                             }
                          });
                     </script>
                 </section>
